refactor(sampling): isolate executable inspection
Move diagnostic formatting into its own translation unit so consumers that do not inspect plans do not link it. Use exhaustive overloaded visitors and clarify the provenance range types.\n\nAssisted-by: Codex (GPT-5) <noreply@openai.com>
